refactor: models/datamodels housed in same folder
