+++ lib/ui/views/survey/shared/toggle_tab_checked.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get/state_manager.dart';
-import 'package:prapare/models/data/survey_tab_data.dart';
+import 'package:prapare/models/survey_tab/survey_tab_data.dart';
 import 'package:prapare/ui/views/survey/survey_controller.dart';
 
 class ToggleTabChecked extends StatelessWidget {
